+++ packages/core/src/components/map-message/map-message.interfaces.ts
@@ -1,0 +1,11 @@
+export interface MapMessage {
+  message: string;
+  variant: "success" | "error" | "instruction";
+  buttonLabels?: string[];
+  dsoActionClick?: (event: CustomEvent<MapMessageActionClickEvent>) => void;
+}
+
+export interface MapMessageActionClickEvent {
+  actionIndex: number;
+  originalEvent: MouseEvent;
+}