+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Task
 * Document Component: types toevoegen ([#3500](https://github.com/dso-toolkit/dso-toolkit/issues/3500))
 * Visueel Regressietesten: Optimaliseer `failureThreshold` ([#3201](https://github.com/dso-toolkit/dso-toolkit/issues/3201))
+* Autosuggest: In Search Bar met state invalid en notFoundLabel ([#3255](https://github.com/dso-toolkit/dso-toolkit/issues/3255))
 
 ## 🌹 Release 89.0.0 - 2026-03-02